+++ functions/src/aggregate/aggregate.algolia.ts
@@ -39,29 +39,4 @@ export class AggregateAlgolia {
       'dataCount:' + dataCount
     );
   }
-
-  /**
-   * algoliaにスキルデータを反映.
-   * スキルデータが更新されるのは、集計後のみなので、firestore.triggerは使わず、replaceAllObjectsでまとめて置換
-   * (algoliaに反映するのは最新データのみ)
-   * @param context
-   */
-  public async replaceSkillData(context: AggregateContext) {
-    console.log('AggregateAlgolia.replaceSkillData.start');
-
-    const algoliaClient = new AlgoliaClinent('skills');
-
-    // algolia用に、objectIDを付加
-    const agSkills: any[] = [];
-    context.skills.forEach((fsSkill) => {
-      const agSkill: any = fsSkill;
-      agSkill.objectID = fsSkill.skillId;
-      agSkills.push(agSkill);
-    });
-
-    // replaceAllにて、全データ一括置換
-    await algoliaClient.getIndex().replaceAllObjects(agSkills);
-
-    console.log('AggregateAlgolia.replaceSkillData.end');
-  }
 }
